Fixed file path so it's pulled from the class path and changed both inputs to doubles instead of ints.

+++ src/main/java/com/palarran/kitesizer/KiteSizeRecommendation.java
@@ -70,7 +70,7 @@ public class KiteSizeRecommendation {
      * @param windSpeed
      * @return true if this wind value is inside the boundaries of this recommendation
      */
-    public boolean windMatches(int windSpeed) {
+    public boolean windMatches(double windSpeed) {
         return windSpeed >= lowerWindSpeed && windSpeed < upperWindSpeed;
     }
